Better level gen
wall blocks are now more likely to spawn and the number of computer inputs in the level testing was increased from 500 to 1000. All this changes make the game a little bit slower but the aprox. loading time is now 1 - 3 sec which is still ok.

However the code will simply stop working after about 5th or 6th level for an unknown reason. No errors detected.