Custom Level Sizes and Other Minor Changes
The game can now generate custom sized levels. Because of this all variables for storing the generated level have been replaced with two arrays: layers and layers_mod. The rest of the code was also modified to support the change.

The logo now displays 0.15 seconds faster.

The game is now extremely vulnerable to crashes mentioned in the previous commits. The reason for them is still unknown but they appear faster in bigger levels.